refactor: replace sprint files from utils/sprint-helpers to components/sprint
